Refactor into some DDD-like architecture
